+++ glib/gstring.c
@@ -446,7 +446,7 @@ g_string_new (const gchar *init)
     string = g_string_sized_new (2);
   else 
     {
-      gint len;
+      gsize len;
 
       len = strlen (init);
       string = g_string_sized_new (len + 2);
@@ -704,7 +704,7 @@ g_string_insert_len (GString     *string,
   if (pos < 0)
     pos = string->len;
   else
-    g_return_val_if_fail (pos <= string->len, string);
+    g_return_val_if_fail ((gsize) pos <= string->len, string);
 
   /* Check whether val represents a substring of string.  This test
      probably violates chapter and verse of the C standards, since
@@ -720,18 +720,18 @@ g_string_insert_len (GString     *string,
       /* At this point, val is valid again.  */
 
       /* Open up space where we are going to insert.  */
-      if (pos < string->len)
+      if ((gsize) pos < string->len)
         g_memmove (string->str + pos + len, string->str + pos, string->len - pos);
 
       /* Move the source part before the gap, if any.  */
-      if (offset < pos)
+      if (offset < (gsize) pos)
         {
-          precount = MIN (len, pos - offset);
+          precount = MIN ((gsize) len, pos - offset);
           memcpy (string->str + pos, val, precount);
         }
 
       /* Move the source part after the gap, if any.  */
-      if (len > precount)
+      if ((gsize) len > precount)
         memcpy (string->str + pos + precount,
                 val + /* Already moved: */ precount + /* Space opened up: */ len,
                 len - precount);
@@ -743,7 +743,7 @@ g_string_insert_len (GString     *string,
       /* If we aren't appending at the end, move a hunk
        * of the old string to the end, opening up space
        */
-      if (pos < string->len)
+      if ((gsize) pos < string->len)
         g_memmove (string->str + pos + len, string->str + pos, string->len - pos);
 
       /* insert the new string */
@@ -1032,7 +1032,7 @@ g_string_insert (GString     *string,
   g_return_val_if_fail (string != NULL, NULL);
   g_return_val_if_fail (val != NULL, string);
   if (pos >= 0)
-    g_return_val_if_fail (pos <= string->len, string);
+    g_return_val_if_fail ((gsize) pos <= string->len, string);
   
   return g_string_insert_len (string, pos, val, -1);
 }
@@ -1059,10 +1059,10 @@ g_string_insert_c (GString *string,
   if (pos < 0)
     pos = string->len;
   else
-    g_return_val_if_fail (pos <= string->len, string);
+    g_return_val_if_fail ((gsize) pos <= string->len, string);
   
   /* If not just an append, move the old stuff */
-  if (pos < string->len)
+  if ((gsize) pos < string->len)
     g_memmove (string->str + pos + 1, string->str + pos, string->len - pos);
 
   string->str[pos] = c;
@@ -1134,10 +1134,10 @@ g_string_insert_unichar (GString *string,
   if (pos < 0)
     pos = string->len;
   else
-    g_return_val_if_fail (pos <= string->len, string);
+    g_return_val_if_fail ((gsize) pos <= string->len, string);
 
   /* If not just an append, move the old stuff */
-  if (pos < string->len)
+  if ((gsize) pos < string->len)
     g_memmove (string->str + pos + charlen, string->str + pos, string->len - pos);
 
   dest = string->str + pos;
@@ -1246,15 +1246,15 @@ g_string_erase (GString *string,
 {
   g_return_val_if_fail (string != NULL, NULL);
   g_return_val_if_fail (pos >= 0, string);
-  g_return_val_if_fail (pos <= string->len, string);
+  g_return_val_if_fail ((gsize) pos <= string->len, string);
 
   if (len < 0)
     len = string->len - pos;
   else
     {
-      g_return_val_if_fail (pos + len <= string->len, string);
+      g_return_val_if_fail ((gsize) (pos + len) <= string->len, string);
 
-      if (pos + len < string->len)
+      if ((gsize) (pos + len) < string->len)
 	g_memmove (string->str + pos, string->str + pos + len, string->len - (pos + len));
     }
